splitting up utils module into several submodules
