--- tobac/utils.py
@@ -1,896 +1,0 @@
-import logging
-
-
-def column_mask_from2D(mask_2D, cube, z_coord="model_level_number"):
-    """Turn 2D watershedding mask into a 3D mask of selected columns.
-
-    Parameters
-    ----------
-    cube : iris.cube.Cube
-        Data cube.
-
-    mask_2D : iris.cube.Cube
-        2D cube containing mask (int id for tacked volumes 0
-        everywhere else).
-
-    z_coord : str
-        Name of the vertical coordinate in the cube.
-
-    Returns
-    -------
-    mask_2D : iris.cube.Cube
-        3D cube containing columns of 2D mask (int id for tracked
-        volumes, 0 everywhere else).
-    """
-
-    from copy import deepcopy
-
-    mask_3D = deepcopy(cube)
-    mask_3D.rename("segmentation_mask")
-    dim = mask_3D.coord_dims(z_coord)[0]
-    for i in range(len(mask_3D.coord(z_coord).points)):
-        slc = [slice(None)] * len(mask_3D.shape)
-        slc[dim] = slice(i, i + 1)
-        mask_out = mask_3D[slc]
-        mask_3D.data[slc] = mask_2D.core_data()
-    return mask_3D
-
-
-def mask_cube_cell(variable_cube, mask, cell, track):
-    """Mask cube for tracked volume of an individual cell.
-
-    Parameters
-    ----------
-    variable_cube : iris.cube.Cube
-        Unmasked data cube.
-
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tracked volumes, 0 everywhere
-        else).
-
-    cell : int
-        Integer id of cell to create masked cube for.
-
-    track : pandas.DataFrame
-        Output of the linking.
-
-    Returns
-    -------
-    variable_cube_out : iris.cube.Cube
-        Masked cube with data for respective cell.
-    """
-
-    from copy import deepcopy
-
-    variable_cube_out = deepcopy(variable_cube)
-    feature_ids = track.loc[track["cell"] == cell, "feature"].values
-    variable_cube_out = mask_cube_features(variable_cube, mask, feature_ids)
-    return variable_cube_out
-
-
-def mask_cube_all(variable_cube, mask):
-    """Mask cube (iris.cube) for tracked volume.
-
-    Parameters
-    ----------
-    variable_cube : iris.cube.Cube
-        Unmasked data cube.
-
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes 0 everywhere
-        else).
-
-    Returns
-    -------
-    variable_cube_out : iris.cube.Cube
-        Masked cube for untracked volume.
-    """
-
-    from dask.array import ma
-    from copy import deepcopy
-
-    variable_cube_out = deepcopy(variable_cube)
-    variable_cube_out.data = ma.masked_where(
-        mask.core_data() == 0, variable_cube_out.core_data()
-    )
-    return variable_cube_out
-
-
-def mask_cube_untracked(variable_cube, mask):
-    """Mask cube (iris.cube) for untracked volume.
-
-    Parameters
-    ----------
-    variable_cube : iris.cube.Cube
-        Unmasked data cube.
-
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes 0 everywhere
-        else).
-
-    Returns
-    -------
-    variable_cube_out : iris.cube.Cube
-        Masked cube for untracked volume.
-    """
-
-    from dask.array import ma
-    from copy import deepcopy
-
-    variable_cube_out = deepcopy(variable_cube)
-    variable_cube_out.data = ma.masked_where(
-        mask.core_data() > 0, variable_cube_out.core_data()
-    )
-    return variable_cube_out
-
-
-def mask_cube(cube_in, mask):
-    """Mask cube where mask is not zero.
-
-    Parameters
-    ----------
-    cube_in : iris.cube.Cube
-        Unmasked data cube.
-
-    mask : iris.cube.Cube
-        Mask to use for masking, >0 where cube is supposed to be masked.
-
-    Returns
-    -------
-    variable_cube_out : iris.cube.Cube
-        Masked cube.
-    """
-
-    from dask.array import ma
-    from copy import deepcopy
-
-    cube_out = deepcopy(cube_in)
-    cube_out.data = ma.masked_where(mask.core_data() != 0, cube_in.core_data())
-    return cube_out
-
-
-def mask_cell(mask, cell, track, masked=False):
-    """Create mask for specific cell.
-
-    Parameters
-    ----------
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tracked volumes 0 everywhere
-        else).
-
-    cell : int
-        Integer id of cell to create masked cube for.
-
-    track : pandas.DataFrame
-        Output of the linking.
-
-    masked : bool, optional
-        Bool determining whether to mask the mask for the cell where
-        it is 0. Default is False.
-
-    Returns
-    -------
-    mask_i : numpy.ndarray
-        Mask for a specific cell.
-    """
-
-    feature_ids = track.loc[track["cell"] == cell, "feature"].values
-    mask_i = mask_features(mask, feature_ids, masked=masked)
-    return mask_i
-
-
-def mask_cell_surface(mask, cell, track, masked=False, z_coord="model_level_number"):
-    """Create surface projection of 3d-mask for individual cell by
-    collapsing one coordinate.
-
-    Parameters
-    ----------
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes, 0 everywhere
-        else).
-
-    cell : int
-        Integer id of cell to create masked cube for.
-
-    track : pandas.DataFrame
-        Output of the linking.
-
-    masked : bool, optional
-        Bool determining whether to mask the mask for the cell where
-        it is 0. Default is False.
-
-    z_coord : str, optional
-        Name of the coordinate to collapse. Default is 'model_level_number'.
-
-    Returns
-    -------
-    mask_i_surface : iris.cube.Cube
-        Collapsed Masked cube for the cell with the maximum value
-        along the collapsed coordinate.
-
-    """
-
-    feature_ids = track.loc[track["cell"] == cell, "feature"].values
-    mask_i_surface = mask_features_surface(
-        mask, feature_ids, masked=masked, z_coord=z_coord
-    )
-    return mask_i_surface
-
-
-def mask_cell_columns(mask, cell, track, masked=False, z_coord="model_level_number"):
-    """Create mask with entire columns for individual cell.
-
-    Parameters
-    ----------
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes 0 everywhere
-        else).
-
-    cell : int
-        Interger id of cell to create the masked cube for.
-
-    track : pandas.DataFrame
-        Output of the linking.
-
-    masked : bool, optional
-        Bool determining whether to mask the mask for the cell where
-        it is 0. Default is False.
-
-    z_coord : str, optional
-        Default is 'model_level_number'.
-
-    Returns
-    -------
-    mask_i : iris.cube.Cube
-        Masked cube for untracked volume.
-
-    Notes
-    -------
-    Function is not working since mask_features_columns()
-    is commented out
-    """
-
-    raise NotImplementedError(
-        "The function mask_cell_columns() is not implemented currently."
-    )
-
-    feature_ids = track.loc[track["cell"] == cell].loc["feature"]
-    mask_i = mask_features_columns(mask, feature_ids, masked=masked, z_coord=z_coord)
-    return mask_i
-
-
-def mask_cube_features(variable_cube, mask, feature_ids):
-    """Mask cube for tracked volume of one or more specific
-    features.
-
-    Parameters
-    ----------
-    variable_cube : iris.cube.Cube
-        Unmasked data cube.
-
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes, 0 everywhere
-        else).
-
-    feature_ids : int or list of ints
-        Integer ids of features to create masked cube for.
-
-    Returns
-    -------
-    variable_cube_out : iris.cube.Cube
-        Masked cube with data for respective features.
-    """
-
-    from dask.array import ma, isin
-    from copy import deepcopy
-
-    variable_cube_out = deepcopy(variable_cube)
-    variable_cube_out.data = ma.masked_where(
-        ~isin(mask.core_data(), feature_ids), variable_cube_out.core_data()
-    )
-    return variable_cube_out
-
-
-def mask_features(mask, feature_ids, masked=False):
-    """Create mask for specific features.
-
-    Parameters
-    ----------
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes 0 everywhere
-        else).
-
-    feature_ids : int or list of ints
-        Integer ids of the features to create the masked cube for.
-
-    masked : bool, optional
-        Bool determining whether to mask the mask for the cell where
-        it is 0. Default is False.
-
-    Returns
-    -------
-    mask_i : numpy.ndarray
-        Masked cube for specific features.
-    """
-
-    from dask.array import ma, isin
-    from copy import deepcopy
-
-    mask_i = deepcopy(mask)
-    mask_i_data = mask_i.core_data()
-    mask_i_data[~isin(mask_i.core_data(), feature_ids)] = 0
-    if masked:
-        mask_i.data = ma.masked_equal(mask_i.core_data(), 0)
-    return mask_i
-
-
-def mask_features_surface(
-    mask, feature_ids, masked=False, z_coord="model_level_number"
-):
-    """Create surface projection of 3d-mask for specific features
-    by collapsing one coordinate.
-
-    Parameters
-    ----------
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes 0 everywhere
-        else).
-
-    feature_ids : int or list of ints
-        Integer ids of the features to create the masked cube for.
-
-    masked : bool, optional
-        Bool determining whether to mask the mask for the cell where
-        it is 0. Default is False
-
-    z_coord : str, optional
-        Name of the coordinate to collapse. Default is
-        'model_level_number'.
-
-    Returns
-    -------
-    mask_i_surface : iris.cube.Cube
-        Collapsed Masked cube for the features with the maximum value
-        along the collapsed coordinate.
-    """
-
-    from iris.analysis import MAX
-    from dask.array import ma, isin
-    from copy import deepcopy
-
-    mask_i = deepcopy(mask)
-    #     mask_i.data=[~isin(mask_i.data,feature_ids)]=0
-    mask_i_data = mask_i.core_data()
-    mask_i_data[~isin(mask_i.core_data(), feature_ids)] = 0
-    mask_i_surface = mask_i.collapsed(z_coord, MAX)
-    if masked:
-        mask_i_surface.data = ma.masked_equal(mask_i_surface.core_data(), 0)
-    return mask_i_surface
-
-
-def mask_all_surface(mask, masked=False, z_coord="model_level_number"):
-    """Create surface projection of 3d-mask for all features
-    by collapsing one coordinate.
-
-    Parameters
-    ----------
-    mask : iris.cube.Cube
-        Cube containing mask (int id for tacked volumes 0 everywhere
-        else).
-
-    masked : bool, optional
-        Bool determining whether to mask the mask for the cell where
-        it is 0. Default is False
-
-    z_coord : str, optional
-        Name of the coordinate to collapse. Default is
-        'model_level_number'.
-
-    Returns
-    -------
-    mask_i_surface : iris.cube.Cube (2D)
-        Collapsed Masked cube for the features with the maximum value
-        along the collapsed coordinate.
-    """
-
-    from iris.analysis import MAX
-    from dask.array import ma, isin
-    from copy import deepcopy
-
-    mask_i = deepcopy(mask)
-    mask_i_surface = mask_i.collapsed(z_coord, MAX)
-    mask_i_surface_data = mask_i_surface.core_data()
-    mask_i_surface.data[mask_i_surface_data > 0] = 1
-    if masked:
-        mask_i_surface.data = ma.masked_equal(mask_i_surface.core_data(), 0)
-    return mask_i_surface
-
-
-# def mask_features_columns(mask,feature_ids,masked=False,z_coord='model_level_number'):
-#     ''' Mask cube for untracked volume
-#     Input:
-#     variable_cube:     iris.cube.Cube
-#                        unmasked data cube
-#     mask:              iris.cube.Cube
-#                        cube containing mask (int id for tacked volumes 0 everywhere else)
-#     Output:
-#     variable_cube_out: iris.cube.Cube
-#                        Masked cube for untracked volume
-#     '''
-#     from iris.analysis import MAX
-#     import numpy as np
-#     from copy import deepcopy
-#     mask_i=deepcopy(mask)
-#     mask_i.data[~np.isin(mask_i.data,feature_ids)]=0
-#     mask_i_surface=mask_i.collapsed(z_coord,MAX)
-#     for cube_slice in mask_i.slices(['time','x','y']):
-#             cube_slice.data=mask_i_surface.core_data()
-#     if masked:
-#         mask_i.data=np.ma.array(mask_i.data,mask=mask_i.data)
-#     return mask_i
-
-
-# def constraint_cell(track,mask_cell,width=None,x=None,):
-#     from iris import Constraint
-#     import numpy as np
-#
-#     time_coord=mask_cell.coord('time')
-#     time_units=time_coord.units
-#
-#     def time_condition(cell):
-#         return time_units.num2date(track.head(n=1)['time']) <= cell <= time_units.num2date(track.tail(n=1)['time'])
-#
-#     constraint_time=Constraint(time=time_condition)
-##     mask_cell_i=mask_cell.extract(constraint_time)
-#     mask_cell_surface_i=mask_cell_surface.extract(constraint_time)
-#
-#     x_dim=mask_cell_surface_i.coord_dims('projection_x_coordinate')[0]
-#     y_dim=mask_cell_surface_i.coord_dims('projection_y_coordinate')[0]
-#     x_coord=mask_cell_surface_i.coord('projection_x_coordinate')
-#     y_coord=mask_cell_surface_i.coord('projection_y_coordinate')
-#
-#     if (mask_cell_surface_i.core_data()>0).any():
-#         box_mask_i=get_bounding_box(mask_cell_surface_i.core_data(),buffer=1)
-#
-#         box_mask=[[x_coord.points[box_mask_i[x_dim][0]],x_coord.points[box_mask_i[x_dim][1]]],
-#                  [y_coord.points[box_mask_i[y_dim][0]],y_coord.points[box_mask_i[y_dim][1]]]]
-#     else:
-#         box_mask=[[np.nan,np.nan],[np.nan,np.nan]]
-#
-#         x_min=box_mask[0][0]
-#         x_max=box_mask[0][1]
-#         y_min=box_mask[1][0]
-#         y_max=box_mask[1][1]
-#     constraint_x=Constraint(projection_x_coordinate=lambda cell: int(x_min) < cell < int(x_max))
-#     constraint_y=Constraint(projection_y_coordinate=lambda cell: int(y_min) < cell < int(y_max))
-#
-#     constraint=constraint_time & constraint_x & constraint_y
-#     return constraint
-
-
-def add_coordinates(t, variable_cube):
-    """Add coordinates from the input cube of the feature detection
-    to the trajectories/features.
-
-    Parameters
-    ----------
-    t : pandas.DataFrame
-        Trajectories/features from feature detection or linking step.
-
-    variable_cube : iris.cube.Cube
-        Input data used for the tracking with coordinate information
-        to transfer to the resulting DataFrame. Needs to contain the
-        coordinate 'time'.
-
-    Returns
-    -------
-    t : pandas.DataFrame
-        Trajectories with added coordinates.
-
-    """
-
-    import numpy as np
-    from scipy.interpolate import interp2d, interp1d
-
-    logging.debug("start adding coordinates from cube")
-
-    # pull time as datetime object and timestr from input data and add it to DataFrame:
-    t["time"] = None
-    t["timestr"] = None
-
-    logging.debug("adding time coordinate")
-
-    time_in = variable_cube.coord("time")
-    time_in_datetime = time_in.units.num2date(time_in.points)
-
-    t["time"] = time_in_datetime[t["frame"]]
-    t["timestr"] = [
-        x.strftime("%Y-%m-%d %H:%M:%S") for x in time_in_datetime[t["frame"]]
-    ]
-
-    # Get list of all coordinates in input cube except for time (already treated):
-    coord_names = [coord.name() for coord in variable_cube.coords()]
-    coord_names.remove("time")
-
-    logging.debug("time coordinate added")
-
-    # chose right dimension for horizontal axis based on time dimension:
-    ndim_time = variable_cube.coord_dims("time")[0]
-    if ndim_time == 0:
-        hdim_1 = 1
-        hdim_2 = 2
-    elif ndim_time == 1:
-        hdim_1 = 0
-        hdim_2 = 2
-    elif ndim_time == 2:
-        hdim_1 = 0
-        hdim_2 = 1
-
-    # create vectors to use to interpolate from pixels to coordinates
-    dimvec_1 = np.arange(variable_cube.shape[hdim_1])
-    dimvec_2 = np.arange(variable_cube.shape[hdim_2])
-
-    # loop over coordinates in input data:
-    for coord in coord_names:
-        logging.debug("adding coord: " + coord)
-        # interpolate 2D coordinates:
-        if variable_cube.coord(coord).ndim == 1:
-
-            if variable_cube.coord_dims(coord) == (hdim_1,):
-                f = interp1d(
-                    dimvec_1,
-                    variable_cube.coord(coord).points,
-                    fill_value="extrapolate",
-                )
-                coordinate_points = f(t["hdim_1"])
-
-            if variable_cube.coord_dims(coord) == (hdim_2,):
-                f = interp1d(
-                    dimvec_2,
-                    variable_cube.coord(coord).points,
-                    fill_value="extrapolate",
-                )
-                coordinate_points = f(t["hdim_2"])
-
-        # interpolate 2D coordinates:
-        elif variable_cube.coord(coord).ndim == 2:
-
-            if variable_cube.coord_dims(coord) == (hdim_1, hdim_2):
-                f = interp2d(dimvec_2, dimvec_1, variable_cube.coord(coord).points)
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_2"], t["hdim_1"])]
-
-            if variable_cube.coord_dims(coord) == (hdim_2, hdim_1):
-                f = interp2d(dimvec_1, dimvec_2, variable_cube.coord(coord).points)
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_1"], t["hdim_2"])]
-
-        # interpolate 3D coordinates:
-        # mainly workaround for wrf latitude and longitude (to be fixed in future)
-
-        elif variable_cube.coord(coord).ndim == 3:
-
-            if variable_cube.coord_dims(coord) == (ndim_time, hdim_1, hdim_2):
-                f = interp2d(
-                    dimvec_2, dimvec_1, variable_cube[0, :, :].coord(coord).points
-                )
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_2"], t["hdim_1"])]
-
-            if variable_cube.coord_dims(coord) == (ndim_time, hdim_2, hdim_1):
-                f = interp2d(
-                    dimvec_1, dimvec_2, variable_cube[0, :, :].coord(coord).points
-                )
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_1"], t["hdim_2"])]
-
-            if variable_cube.coord_dims(coord) == (hdim_1, ndim_time, hdim_2):
-                f = interp2d(
-                    dimvec_2, dimvec_1, variable_cube[:, 0, :].coord(coord).points
-                )
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_2"], t["hdim_1"])]
-
-            if variable_cube.coord_dims(coord) == (hdim_1, hdim_2, ndim_time):
-                f = interp2d(
-                    dimvec_2, dimvec_1, variable_cube[:, :, 0].coord(coord).points
-                )
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_2"], t["hdim1"])]
-
-            if variable_cube.coord_dims(coord) == (hdim_2, ndim_time, hdim_1):
-                f = interp2d(
-                    dimvec_1, dimvec_2, variable_cube[:, 0, :].coord(coord).points
-                )
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_1"], t["hdim_2"])]
-
-            if variable_cube.coord_dims(coord) == (hdim_2, hdim_1, ndim_time):
-                f = interp2d(
-                    dimvec_1, dimvec_2, variable_cube[:, :, 0].coord(coord).points
-                )
-                coordinate_points = [f(a, b) for a, b in zip(t["hdim_1"], t["hdim_2"])]
-
-        # write resulting array or list into DataFrame:
-        t[coord] = coordinate_points
-
-        logging.debug("added coord: " + coord)
-    return t
-
-
-def get_bounding_box(x, buffer=1):
-    """Finds the bounding box of a ndarray, i.e. the smallest
-    bounding rectangle for nonzero values as explained here:
-    https://stackoverflow.com/questions/31400769/bounding-box-of-numpy-array
-
-    Parameters
-    ----------
-    x : numpy.ndarray
-        Array for which the bounding box is to be determined.
-
-    buffer : int, optional
-        Number to set a buffer between the nonzero values and
-        the edges of the box. Default is 1.
-
-    Returns
-    -------
-    bbox : list
-        Dimensionwise list of the indices representing the edges
-        of the bounding box.
-
-    """
-
-    from numpy import delete, arange, diff, nonzero, array
-
-    mask = x == 0
-
-    bbox = []
-    all_axis = arange(x.ndim)
-    # loop over dimensions
-    for kdim in all_axis:
-        nk_dim = delete(all_axis, kdim)
-        mask_i = mask.all(axis=tuple(nk_dim))
-        dmask_i = diff(mask_i)
-        idx_i = nonzero(dmask_i)[0]
-        # for case where there is no value in idx_i
-        if len(idx_i) == 0:
-            idx_i = array([0, x.shape[kdim] - 1])
-        # for case where there is only one value in idx_i
-        elif len(idx_i) == 1:
-            idx_i = array([idx_i, idx_i])
-        # make sure there is two values in idx_i
-        elif len(idx_i) > 2:
-            idx_i = array([idx_i[0], idx_i[-1]])
-        # caluclate min and max values for idx_i and append them to list
-        idx_min = max(0, idx_i[0] + 1 - buffer)
-        idx_max = min(x.shape[kdim] - 1, idx_i[1] + 1 + buffer)
-        bbox.append([idx_min, idx_max])
-    return bbox
-
-
-def get_spacings(field_in, grid_spacing=None, time_spacing=None):
-    """Determine spatial and temporal grid spacing of the
-    input data.
-
-    Parameters
-    ----------
-    field_in : iris.cube.Cube
-        Input field where to get spacings.
-
-    grid_spacing : float, optional
-        Manually sets the grid spacing if specified.
-        Default is None.
-
-    time_spacing : float, optional
-        Manually sets the time spacing if specified.
-        Default is None.
-
-    Returns
-    -------
-    dxy : float
-        Grid spacing in metres.
-
-    dt : float
-        Time resolution in seconds.
-
-    Raises
-    ------
-    ValueError
-        If input_cube does not contain projection_x_coord and
-        projection_y_coord or keyword argument grid_spacing.
-
-    """
-
-    import numpy as np
-    from copy import deepcopy
-
-    # set horizontal grid spacing of input data
-    # If cartesian x and y corrdinates are present, use these to determine dxy (vertical grid spacing used to transfer pixel distances to real distances):
-    coord_names = [coord.name() for coord in field_in.coords()]
-
-    if (
-        "projection_x_coordinate" in coord_names
-        and "projection_y_coordinate" in coord_names
-    ) and (grid_spacing is None):
-        x_coord = deepcopy(field_in.coord("projection_x_coordinate"))
-        x_coord.convert_units("metre")
-        dx = np.diff(field_in.coord("projection_y_coordinate")[0:2].points)[0]
-        y_coord = deepcopy(field_in.coord("projection_y_coordinate"))
-        y_coord.convert_units("metre")
-        dy = np.diff(field_in.coord("projection_y_coordinate")[0:2].points)[0]
-        dxy = 0.5 * (dx + dy)
-    elif grid_spacing is not None:
-        dxy = grid_spacing
-    else:
-        raise ValueError(
-            "no information about grid spacing, need either input cube with projection_x_coord and projection_y_coord or keyword argument grid_spacing"
-        )
-
-    # set horizontal grid spacing of input data
-    if time_spacing is None:
-        # get time resolution of input data from first to steps of input cube:
-        time_coord = field_in.coord("time")
-        dt = (
-            time_coord.units.num2date(time_coord.points[1])
-            - time_coord.units.num2date(time_coord.points[0])
-        ).seconds
-    elif time_spacing is not None:
-        # use value of time_spacing for dt:
-        dt = time_spacing
-    return dxy, dt
-
-
-def get_label_props_in_dict(labels):
-    """Function to get the label properties into a dictionary format.
-
-    Parameters
-    ----------
-    labels : 2D array-like
-        Output of the `skimage.measure.label` function.
-
-    Returns
-    -------
-    region_properties_dict: dict
-        Output from skimage.measure.regionprops in dictionary
-        format, where they key is the label number.
-    """
-
-    import skimage.measure
-
-    region_properties_raw = skimage.measure.regionprops(labels)
-    region_properties_dict = {
-        region_prop.label: region_prop for region_prop in region_properties_raw
-    }
-
-    return region_properties_dict
-
-
-def get_indices_of_labels_from_reg_prop_dict(region_property_dict):
-    """Function to get the x and y indices (as well as point count) of
-    all labeled regions.
-
-    Parameters
-    ----------
-    region_property_dict : dict of region_property objects
-        This dict should come from the get_label_props_in_dict function.
-
-    Returns
-    -------
-    curr_loc_indices : dict
-        The number of points in the label number (key: label number).
-
-    y_indices : dict
-        The y indices in the label number (key: label number).
-
-    x_indices : dict
-        The x indices in the label number (key: label number).
-
-    Raises
-    ------
-    ValueError
-        A ValueError is raised if there are no regions in the region
-        property dict.
-    """
-
-    import numpy as np
-
-    if len(region_property_dict) == 0:
-        raise ValueError("No regions!")
-
-    y_indices = dict()
-    x_indices = dict()
-    curr_loc_indices = dict()
-
-    # loop through all skimage identified regions
-    for region_prop_key in region_property_dict:
-        region_prop = region_property_dict[region_prop_key]
-        index = region_prop.label
-        curr_y_ixs, curr_x_ixs = np.transpose(region_prop.coords)
-
-        y_indices[index] = curr_y_ixs
-        x_indices[index] = curr_x_ixs
-        curr_loc_indices[index] = len(curr_y_ixs)
-
-    return (curr_loc_indices, y_indices, x_indices)
-
-
-def spectral_filtering(
-    dxy, field_in, lambda_min, lambda_max, return_transfer_function=False
-):
-    """This function creates and applies a 2D transfer function that
-    can be used as a bandpass filter to remove certain wavelengths
-    of an atmospheric input field (e.g. vorticity, IVT, etc).
-
-    Parameters:
-    -----------
-    dxy : float
-        Grid spacing in m.
-
-    field_in: numpy.array
-        2D field with input data.
-
-    lambda_min: float
-        Minimum wavelength in m.
-
-    lambda_max: float
-        Maximum wavelength in m.
-
-    return_transfer_function: boolean, optional
-        default: False. If set to True, then the 2D transfer function and
-        the corresponding wavelengths are returned.
-
-    Returns:
-    --------
-    filtered_field: numpy.array
-        Spectrally filtered 2D field of data (with same shape as input data).
-
-    transfer_function: tuple
-        Two 2D fields, where the first one corresponds to the wavelengths
-        in the spectral space of the domain and the second one to the 2D
-        transfer function of the bandpass filter. Only returned, if
-        return_transfer_function is True.
-    """
-
-    import numpy as np
-    from scipy import signal
-    from scipy import fft
-
-    # check if valid value for dxy is given
-    if dxy <= 0:
-        raise ValueError(
-            "Invalid value for dxy. Please provide the grid spacing in meter."
-        )
-
-    # get number of grid cells in x and y direction
-    Ni = field_in.shape[-2]
-    Nj = field_in.shape[-1]
-    # wavenumber space
-    m, n = np.meshgrid(np.arange(Ni), np.arange(Nj), indexing="ij")
-
-    # if domain is squared:
-    if Ni == Nj:
-        wavenumber = np.sqrt(m**2 + n**2)
-        lambda_mn = (2 * Ni * (dxy)) / wavenumber
-    else:
-        # if domain is a rectangle:
-        # alpha is the normalized wavenumber in wavenumber space
-        alpha = np.sqrt(m**2 / Ni**2 + n**2 / Nj**2)
-        # compute wavelengths for target grid in m
-        lambda_mn = 2 * dxy / alpha
-
-    ############### create a 2D bandpass filter (butterworth) #######################
-    b, a = signal.iirfilter(
-        2,
-        [1 / lambda_max, 1 / lambda_min],
-        btype="band",
-        ftype="butter",
-        fs=1 / dxy,
-        output="ba",
-    )
-    w, h = signal.freqz(b, a, 1 / lambda_mn.flatten(), fs=1 / dxy)
-    transfer_function = np.reshape(abs(h), lambda_mn.shape)
-
-    # 2-dimensional discrete cosine transformation to convert data to spectral space
-    spectral = fft.dctn(field_in.data)
-    # multiplication of spectral coefficients with transfer function
-    filtered = spectral * transfer_function
-    # inverse discrete cosine transformation to go back from spectral to original space
-    filtered_field = fft.idctn(filtered)
-
-    if return_transfer_function is True:
-        return (lambda_mn, transfer_function), filtered_field
-    else:
-        return filtered_field
